Move Search Area setting to separate file
