+++ src/main/java/net/wurstclient/hacks/SearchHack.java
@@ -45,12 +45,12 @@ import net.minecraft.util.math.ChunkPos;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Matrix4f;
 import net.minecraft.world.chunk.Chunk;
-import net.minecraft.world.chunk.EmptyChunk;
 import net.wurstclient.Category;
 import net.wurstclient.events.PacketInputListener;
 import net.wurstclient.events.RenderListener;
 import net.wurstclient.events.UpdateListener;
 import net.wurstclient.hack.Hack;
+import net.wurstclient.hacks.search.SearchArea;
 import net.wurstclient.settings.BlockSetting;
 import net.wurstclient.settings.EnumSetting;
 import net.wurstclient.settings.SliderSetting;
@@ -68,10 +68,10 @@ public final class SearchHack extends Hack
 	private final BlockSetting block = new BlockSetting("Block",
 		"The type of block to search for.", "minecraft:diamond_ore", false);
 	
-	private final EnumSetting<Area> area = new EnumSetting<>("Area",
+	private final EnumSetting<SearchArea> area = new EnumSetting<>("Area",
 		"The area around the player to search in.\n"
 			+ "Higher values require a faster computer.",
-		Area.values(), Area.D11);
+		SearchArea.values(), SearchArea.D11);
 	
 	private final SliderSetting limit = new SliderSetting("Limit",
 		"The maximum number of blocks to display.\n"
@@ -184,11 +184,10 @@ public final class SearchHack extends Hack
 		BlockPos eyesPos = new BlockPos(RotationUtils.getEyesPos());
 		
 		ChunkPos center = MC.player.getChunkPos();
-		int range = area.getSelected().chunkRange;
 		int dimensionId = MC.world.getRegistryKey().toString().hashCode();
 		
-		addSearchersInRange(center, range, currentBlock, dimensionId);
-		removeSearchersOutOfRange(center, range);
+		addSearchersInRange(center, currentBlock, dimensionId);
+		removeSearchersOutOfRange(center);
 		replaceSearchersWithDifferences(currentBlock, dimensionId);
 		replaceSearchersWithChunkUpdate(currentBlock, dimensionId);
 		
@@ -254,10 +253,11 @@ public final class SearchHack extends Hack
 		GL11.glDisable(GL11.GL_LINE_SMOOTH);
 	}
 	
-	private void addSearchersInRange(ChunkPos center, int chunkRange,
-		Block block, int dimensionId)
+	private void addSearchersInRange(ChunkPos center, Block block,
+		int dimensionId)
 	{
-		ArrayList<Chunk> chunksInRange = getChunksInRange(center, chunkRange);
+		ArrayList<Chunk> chunksInRange =
+			area.getSelected().getChunksInRange(center);
 		
 		for(Chunk chunk : chunksInRange)
 		{
@@ -268,31 +268,12 @@ public final class SearchHack extends Hack
 		}
 	}
 	
-	private ArrayList<Chunk> getChunksInRange(ChunkPos center, int chunkRange)
-	{
-		ArrayList<Chunk> chunksInRange = new ArrayList<>();
-		
-		for(int x = center.x - chunkRange; x <= center.x + chunkRange; x++)
-			for(int z = center.z - chunkRange; z <= center.z + chunkRange; z++)
-			{
-				Chunk chunk = MC.world.getChunk(x, z);
-				if(chunk instanceof EmptyChunk)
-					continue;
-				
-				chunksInRange.add(chunk);
-			}
-		
-		return chunksInRange;
-	}
-	
-	private void removeSearchersOutOfRange(ChunkPos center, int chunkRange)
+	private void removeSearchersOutOfRange(ChunkPos center)
 	{
 		for(ChunkSearcher searcher : new ArrayList<>(searchers.values()))
 		{
 			ChunkPos searcherPos = searcher.getChunk().getPos();
-			
-			if(Math.abs(searcherPos.x - center.x) <= chunkRange
-				&& Math.abs(searcherPos.z - center.z) <= chunkRange)
+			if(area.getSelected().isInRange(searcherPos, center))
 				continue;
 			
 			removeSearcher(searcher);
@@ -476,41 +457,6 @@ public final class SearchHack extends Hack
 		}catch(InterruptedException | ExecutionException e)
 		{
 			throw new RuntimeException(e);
-		}
-	}
-	
-	private enum Area
-	{
-		D3("3x3 chunks", 1),
-		D5("5x5 chunks", 2),
-		D7("7x7 chunks", 3),
-		D9("9x9 chunks", 4),
-		D11("11x11 chunks", 5),
-		D13("13x13 chunks", 6),
-		D15("15x15 chunks", 7),
-		D17("17x17 chunks", 8),
-		D19("19x19 chunks", 9),
-		D21("21x21 chunks", 10),
-		D23("23x23 chunks", 11),
-		D25("25x25 chunks", 12),
-		D27("27x27 chunks", 13),
-		D29("29x29 chunks", 14),
-		D31("31x31 chunks", 15),
-		D33("33x33 chunks", 16);
-		
-		private final String name;
-		private final int chunkRange;
-		
-		private Area(String name, int chunkRange)
-		{
-			this.name = name;
-			this.chunkRange = chunkRange;
-		}
-		
-		@Override
-		public String toString()
-		{
-			return name;
 		}
 	}
 }
